Move tests away from `__init__`
